Integrate path approximation into `approx`
