+++ crates/fj-kernel/src/algorithms/approx/mod.rs
@@ -4,6 +4,7 @@ pub mod curve;
 pub mod cycle;
 pub mod edge;
 pub mod face;
+pub mod path;
 pub mod shell;
 pub mod sketch;
 pub mod solid;
